requirements.txt
SpeechRecognition
pyttsx3
pyaudio
google-generativeai
requests
selenium
pywhatkit
